Перенес реализацию в cpp файл
